requirements.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.10.0
djangorestframework==3.12.2
Faker==4.14.2
Pillow==8.4.0
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
validate-docbr==1.8.2
